requirements.txt
django==4.1.12
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==23.3
